Add a new transport "serial" for USB attached UARTs
In the URI, the hostname (mandatory) specifies the serial port settings
in the format of the Linux serial consile options, e.g. "9600n8" (or just
"9600") for 9600bps, 8 bits, no parity.

If the serial device has multiple ports, a port number (counting from 0)
can be added to the connection string, e.g. 9600n8:1.